+++ Editor/DeploymentWindowApi/DeploymentItem.cs
@@ -9,10 +9,11 @@ namespace Unity.Services.DeploymentApi.Editor
     /// <summary>
     /// Concrete implementation of IDeploymentItem. Should be implemented over the interface.
     /// </summary>
-    class DeploymentItem : IDeploymentItem
+    class DeploymentItem : IDeploymentItem, ITypedItem
     {
         protected string m_Name;
         protected string m_Path;
+        protected string m_Type;
         protected float m_Progress;
         protected DeploymentStatus m_Status;
         protected ObservableCollection<AssetState> m_States = new ObservableCollection<AssetState>();
@@ -29,6 +30,12 @@ namespace Unity.Services.DeploymentApi.Editor
         {
             get => m_Path;
             set => SetField(ref m_Path, value);
+        }
+
+        public string Type
+        {
+            get => m_Type;
+            set => SetField(ref m_Type, value);
         }
 
         public virtual float Progress