+++ Editor/DeploymentWindowApi/DeploymentStatus.cs
@@ -12,15 +12,26 @@ namespace Unity.Services.DeploymentApi.Editor
     struct DeploymentStatus
     {
         //n.b.: Do not mark as readonly. Required for reload domain persistence
-        SyncStatus m_SyncStatus;
         string m_Message;
         string m_MessageDetail;
         SeverityLevel m_MessageSeverity;
 
         /// <summary>
-        /// Resource status relative to the remote resource
+        /// A status to represent an item that is up to date with the remote.
         /// </summary>
-        public SyncStatus SyncStatus => m_SyncStatus;
+        public static readonly DeploymentStatus UpToDate = new DeploymentStatus("Up to date", string.Empty, SeverityLevel.Success);
+        /// <summary>
+        /// A status to represent an item that was modified locally.
+        /// </summary>
+        public static readonly DeploymentStatus ModifiedLocally = new DeploymentStatus("Modified locally, deploy to update", string.Empty, SeverityLevel.Warning);
+        /// <summary>
+        ///A status to represent an item that failed to deploy.
+        /// </summary>
+        public static readonly DeploymentStatus FailedToDeploy = new DeploymentStatus("Failed to deploy", string.Empty, SeverityLevel.Error);
+        /// <summary>
+        /// An empty status.
+        /// </summary>
+        public static readonly DeploymentStatus Empty = new DeploymentStatus(string.Empty, string.Empty, SeverityLevel.None);
 
         /// <summary>
         /// Message associated with a deployment result
@@ -40,34 +51,14 @@ namespace Unity.Services.DeploymentApi.Editor
         /// <summary>
         /// Creates a DeploymentStatus struct
         /// </summary>
-        /// <param name="syncStatus">Resource status relative to the remote resource</param>
-        /// <param name="message">The DeploymentStatus message if any</param>
-        /// <param name="messageDetail">Details regarding the message</param>
-        /// <param name="messageSeverity">The severity of the message</param>
-        public DeploymentStatus(
-            SyncStatus syncStatus,
-            string message = null,
-            string messageDetail = null,
-            SeverityLevel messageSeverity = SeverityLevel.Error)
-        {
-            m_SyncStatus = syncStatus;
-            m_Message = message;
-            m_MessageDetail = messageDetail;
-            m_MessageSeverity = messageSeverity;
-        }
-
-        // <summary>
-        /// Creates a DeploymentStatus struct
-        /// </summary>
         /// <param name="message">The DeploymentStatus message if any</param>
         /// <param name="messageDetail">Details regarding the message</param>
         /// <param name="messageSeverity">The severity of the message</param>
         public DeploymentStatus(
             string message = null,
             string messageDetail = null,
-            SeverityLevel messageSeverity = SeverityLevel.Error)
+            SeverityLevel messageSeverity = SeverityLevel.None)
         {
-            m_SyncStatus = SyncStatus.Unknown;
             m_Message = message;
             m_MessageDetail = messageDetail;
             m_MessageSeverity = messageSeverity;