+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this shared code project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.5] - 2022-11-16
+### Added
+- New interface ITypedItem to allow deployment items to specify a sub-type for their service assets.
+- `DeploymentStatus` has new pre-built statuses such as `DeploymentStatus.UpToDate` to facilitate status updates.
+
 ## [1.0.0-pre.3] - 2022-09-07
 ### This is the first release of *Unity Package com.unity.services.deployment.api*.